3.1.0
Fixed some marker icon popping issues.
Improved performance by only doing one pass of the actors in the scene.
Fixed bug that caused that the markers were shown in Dragon's Eye Minimap despite the settings. This also increased performance of that mod.
Removed PostCreateMarkers from the API as now it is not needed.